+++ app/src/main/java/com/eljeff/appfinalproject/ui/cart/CartFragment.kt
@@ -1,10 +1,11 @@
-package com.eljeff.appfinalproject
+package com.eljeff.appfinalproject.ui.cart
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.eljeff.appfinalproject.R
 
 class CartFragment : Fragment() {
