+++ app/src/main/java/com/eljeff/appfinalproject/ui/chat/ChatFragment.kt
@@ -1,18 +1,21 @@
-package com.eljeff.appfinalproject
+package com.eljeff.appfinalproject.ui.chat
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.eljeff.appfinalproject.R
 
-class PurchaseFragment : Fragment() {
+
+class ChatFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_purchase, container, false)
+        return inflater.inflate(R.layout.fragment_chat, container, false)
     }
+
 }
