+++ app/src/main/java/com/eljeff/appfinalproject/ui/profile/ProfileFragment.kt
@@ -1,20 +1,20 @@
-package com.eljeff.appfinalproject
+package com.eljeff.appfinalproject.ui.profile
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.eljeff.appfinalproject.R
 
-
-class ChatFragment : Fragment() {
+class ProfileFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_chat, container, false)
+        return inflater.inflate(R.layout.fragment_profile, container, false)
     }
 
 }
