+++ app/src/main/java/com/eljeff/appfinalproject/ui/purchase/PurchaseFragment.kt
@@ -1,19 +1,19 @@
-package com.eljeff.appfinalproject
+package com.eljeff.appfinalproject.ui.purchase
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.eljeff.appfinalproject.R
 
-class ProfileFragment : Fragment() {
+class PurchaseFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_profile, container, false)
+        return inflater.inflate(R.layout.fragment_purchase, container, false)
     }
-
 }
