+++ app/src/main/res/layout/fragment_cart.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/background"
     android:padding="8dp"
-    tools:context=".CartFragment">
+    tools:context=".ui.cart.CartFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
