+++ app/src/main/res/layout/fragment_chat.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/background"
     android:padding="8dp"
-    tools:context=".ChatFragment">
+    tools:context=".ui.chat.ChatFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
