+++ app/src/main/res/layout/fragment_purchase.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/background"
     android:padding="8dp"
-    tools:context=".PurchaseFragment">
+    tools:context=".ui.purchase.PurchaseFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
